requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
SQLAlchemy==2.0.35
python-docx==1.1.2
python-multipart==0.0.9